Update Markdown editor plugin link
The current link points to an old repo that is incompatible with the latest (1.8.6) q2a.  
(q2a gets unresponsive when trying to open the plugins tab on the admin panel.)  
There is a newer repo (with the same author) that has updated the plugin and fixed the above problem.