Add Repository last updated tags
Note: This is supposed to be an extension for the last PR #91

This PR introduces a new feature that enhances plugin visibility and compatibility display:

- Fetches all plugin repository links and retrieves their metadata.

- Stores metadata locally using localStorage for performance, and also not bombard fetches to `raw.githubusercontent.com` on every page request.

- Dynamically generates a footer section for each plugin with:

  - Repository update date

  - "Q2A Tested With" tag based on the plugin’s `max_q2a` metadata.json key, if available.

This helps users quickly assess which plugins are compatible with their Q2A version and when they were last updated.

I also added a message to the Plugins & Themes pages, explaining that "outdated" plugins doesn't necessarily mean that they don't work. Some of them just don't require to be updated as frequently.

Also, everybody can create PRs to update the `max_q2a` version of a plugin they know it works, because they're using it.